+++ src/Operations/MapAsync.cs
@@ -17,9 +17,6 @@ partial struct Result<TValue>
         => _hasValue ? await map(_value) : _error;
     public Task<Result<TResult>> MapAsync<TResult>(Func<TValue, Task<Result<TResult>>> map)
         => _hasValue ? map(_value) : Task.FromResult(Result.Error<TResult>(_error));
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public async Task<Result<TResult, TError>> MapAsync<TResult, TError>(Func<TValue, Task<TResult>> map, Func<Exception, TError> errorMap)
-        => _hasValue ? await map(_value) : errorMap(_error);
 }
 
 partial struct Result<TValue, TError>
@@ -28,13 +25,6 @@ partial struct Result<TValue, TError>
         => _hasValue ? await map(_value) : _error;
     public Task<Result<TResult, TError>> MapAsync<TResult>(Func<TValue, Task<Result<TResult, TError>>> map)
         => _hasValue ? map(_value) : Task.FromResult(Result.Error<TResult, TError>(_error));
-    [OverloadResolutionPriority(1)]
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public async Task<Result<TResult>> MapAsync<TResult>(Func<TValue, Task<TResult>> map, Func<TError, Exception> errorMap)
-        => _hasValue ? await map(_value) : errorMap(_error);
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public async Task<Result<TResult, TNewError>> MapAsync<TResult, TNewError>(Func<TValue, Task<TResult>> map, Func<TError, TNewError> errorMap)
-        => _hasValue ? await map(_value) : errorMap(_error);
 }
 
 public static class OptionMapAsyncExtensions
@@ -53,39 +43,16 @@ public static class OptionMapAsyncExtensions
         => (await optionTask).Map(map);
     public static async Task<Result<TResult>> MapAsync<TValue, TResult>(this Task<Result<TValue>> optionTask, Func<TValue, Result<TResult>> map)
         => (await optionTask).Map(map);
-
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult, TNewError>> MapAsync<TValue, TResult, TNewError>(this Task<Result<TValue>> optionTask, Func<TValue, TResult> map, Func<Exception, TNewError> errorMap)
-        => (await optionTask).Map(map, errorMap);
     public static async Task<Result<TResult>> MapAsync<TValue, TResult>(this Task<Result<TValue>> resultTask, Func<TValue, Task<TResult>> map)
         => await (await resultTask).MapAsync(map);
     public static async Task<Result<TResult>> MapAsync<TValue, TResult>(this Task<Result<TValue>> resultTask, Func<TValue, Task<Result<TResult>>> map)
         => await (await resultTask).MapAsync(map);
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult, TError>> MapAsync<TValue, TResult, TError>(this Task<Result<TValue>> resultTask, Func<TValue, Task<TResult>> map, Func<Exception, TError> errorMap)
-        => await (await resultTask).MapAsync(map, errorMap);
+
 
     public static async Task<Result<TResult, TError>> MapAsync<TValue, TError, TResult>(this Task<Result<TValue, TError>> optionTask, Func<TValue, TResult> map)
         => (await optionTask).Map(map);
-    
-    [OverloadResolutionPriority(1)]
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult, Exception>> MapAsync<TValue, TError, TResult>(this Task<Result<TValue, TError>> optionTask, Func<TValue, TResult> map, Func<TError, Exception> errorMap)
-        => (await optionTask).Map(map, errorMap);
-
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult, TNewError>> MapAsync<TValue, TError, TResult, TNewError>(this Task<Result<TValue, TError>> optionTask, Func<TValue, TResult> map, Func<TError, TNewError> errorMap)
-        => (await optionTask).Map(map, errorMap);
     public static async Task<Result<TResult, TError>> MapAsync<TValue, TError, TResult>(this Task<Result<TValue, TError>> resultTask, Func<TValue, Task<TResult>> map)
         => await (await resultTask).MapAsync(map);
     public static async Task<Result<TResult, TError>> MapAsync<TValue, TError, TResult>(this Task<Result<TValue, TError>> resultTask, Func<TValue, Task<Result<TResult, TError>>> map)
         => await (await resultTask).MapAsync(map);
-    
-    [OverloadResolutionPriority(1)]
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult>> MapAsync<TValue, TError, TResult>(this Task<Result<TValue, TError>> resultTask, Func<TValue, Task<TResult>> map, Func<TError, Exception> errorMap)
-        => await (await resultTask).MapAsync(map, errorMap);
-    [Obsolete("use MapAsync and MapError", error: true)]
-    public static async Task<Result<TResult, TNewError>> MapAsync<TValue, TError, TResult, TNewError>(this Task<Result<TValue, TError>> resultTask, Func<TValue, Task<TResult>> map, Func<TError, TNewError> errorMap)
-        => await (await resultTask).MapAsync(map, errorMap);
 }